Pré finalização da Compra

+++ Models/CondicaoPgto.cs
@@ -62,8 +62,8 @@ namespace Sistema.Models
             public string dsCondicaoPgto { get; set; }
             public int? idFormaPgto { get; set; }
             public string dsFormaPgto { get; set; }
-            public short? nrParcela { get; set; }
-            public short? qtDias { get; set; }
+            public int? nrParcela { get; set; }
+            public int? qtDias { get; set; }
             public decimal txPercentual { get; set; }
         }
